Fixed betty, compiler errors, added prototypes to header, tested the project
